ms sql server connection

+++ src/Controller/TravelLengthComparator.java
@@ -7,7 +7,7 @@ public class TravelLengthComparator implements Comparator<Travel> {
 
 	@Override
 	public int compare(Travel o1, Travel o2) {
-		return Integer.compare(o1.getLength(), o2.getLength());
+		return Float.compare(o1.getLength(), o2.getLength());
 	}
 
 }
