+++ src/Controller/TravelMileageComparator.java
@@ -7,7 +7,7 @@ public class TravelMileageComparator implements Comparator<Travel> {
 
 	@Override
 	public int compare(Travel o1, Travel o2) {
-		return Integer.compare(o1.getMileage(), o2.getMileage());
+		return Float.compare(o1.getMileage(), o2.getMileage());
 	}
 
 }
